requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
sqlite3==0.4.6
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
cryptography==41.0.8
bcrypt==4.1.2
smtplib-ssl==1.0.4
email-validator==2.1.0
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
websockets==12.0
aioredis==2.0.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
plotly==5.17.0
kiteconnect==4.2.0